% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWidget.R
\name{getDependencies}
\alias{getDependencies}
\title{Get dependency html text format}
\usage{
getDependencies(html, url = "", basePath = "", srcType = c("href",
  "file"), encodeFunc = htmltools::urlEncodePath)
}
\arguments{
\item{html}{HTML content}

\item{url}{url to store dependencies}

\item{basePath}{folder under \code{url} to store dependencies}

\item{srcType}{The type of src paths to use; valid values are \code{file} or
\code{href}.}

\item{encodeFunc}{The function to use to encode the path part of a URL. The
default should generally be used.}
}
\value{
dependencies in HTML text format
}
\description{
Get dependency html text format
}

