% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWidget.R
\name{getDependencyDir}
\alias{getDependencyDir}
\title{Get relative path for dependency}
\usage{
getDependencyDir(dependency, mustWork = TRUE)
}
\arguments{
\item{dependency}{a dependency of html content}

\item{mustWork}{If \code{TRUE} and \code{dependency} does not point to a
directory on disk (but rather a URL location), an error is raised. If
\code{FALSE} then non-disk dependencies are returned without modification.}
}
\value{
relative path corresponding to that dependency
}
\description{
Get relative path for dependency
}

