% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rWidget.R
\name{getDivData}
\alias{getDivData}
\title{Get div and data in html text format}
\usage{
getDivData(html)
}
\arguments{
\item{html}{HTML content}
}
\value{
div and data in html text format
}
\description{
Get div and data in html text format
}

